+++ docs/changelog.rst
@@ -18,6 +18,7 @@ Currently in development.
 - Allow administrators to add users to groups without an invitation
 - Allow disabling project group deletion by non-administrators
 - Added signed export and import for .eln files
+- Added topics to locations
 - Fix handling of quantities with logarithmic units